# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CommandCentre OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaign-plan/milestones:
    get:
      operationId: listMilestones
      tags: [campaign-plan]
      summary: List campaign milestones
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [campaign-plan]
      summary: Create a milestone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /campaign-plan/milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [campaign-plan]
      summary: Update milestone status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [campaign-plan]
      summary: Delete a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /campaign-plan/countdown:
    get:
      operationId: getElectionCountdown
      tags: [campaign-plan]
      summary: Days until election day
      responses:
        "200":
          description: Countdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionCountdown"
    post:
      operationId: setElectionDate
      tags: [campaign-plan]
      summary: Set election date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElectionDateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionCountdown"

  # ─── KOL / Influencer ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Milestone:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        dueDate: { type: string }
        status: { type: string, enum: [pending, in_progress, completed, overdue] }
        category: { type: string }
        owner: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, title, dueDate, status, category, createdAt]
    MilestoneInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        dueDate: { type: string }
        category: { type: string }
        owner: { type: string }
      required: [title, dueDate, category]
    MilestoneUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        dueDate: { type: string }
        status: { type: string, enum: [pending, in_progress, completed, overdue] }
        category: { type: string }
        owner: { type: string }
...
    ElectionCountdown:
      type: object
      properties:
        electionDate: { type: ["string", "null"] }
        daysRemaining: { type: ["integer", "null"] }
        phase: { type: string }
      required: [phase]
    ElectionDateInput:
      type: object
      properties:
        electionDate: { type: string }
      required: [electionDate]

    # ─── KOL ─────────────────────────────────────────────────────────────
...